requirements.txt
asgiref==3.3.4
autopep8==1.5.7
Django==3.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
gpx-converter==1.7.3
gpxpy==1.4.2
gunicorn==20.1.0
numpy==1.20.3
pandas==1.2.4
pycodestyle==2.7.0
Pygments==2.9.0
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
